Testuser1

add new lib's func for ken

+++ src/Food/Library/Album/AlbumLibrary.php
@@ -12,25 +12,31 @@ class AlbumLibrary extends Seed
     /**
      * 取得所有album資料
      *
-     * @return array $allalbum
-     *         [
-     *         Title => Title,
-     *         Description => Description,
-     *         Token => Token
-     *         ];
+    * @return array $allalbum 
+    *     [Token =>[ 
+    *         Title => Title,
+    *         Description => Description
+     *     ]];
+    *         or array $end = [ending => 'no albums']
      */
     public static function allalbum()
     {
         $albums = Album::listAll();
-        foreach ($albums as $a) {
-            $album = Album::load($a);
-            $allalbum[$a] = [
-                'Title' => $album->getTitle(),
-                'Description' => $album->getDescription(),
-                'Token' => $album->getToken()
+        if ($albums != null) {
+            foreach ($albums as $a) {
+                $album = Album::load($a);
+                $allalbum[$a] = [
+                    'Title' => $album->getTitle(),
+                    'Description' => $album->getDescription()
+                ];
+            }
+            return $allalbum;
+        } else {
+            $end = [
+                "ending" => "no albums"
             ];
+            return $end;
         }
-        return $allalbum;
     }
 
     /**
@@ -39,39 +45,81 @@ class AlbumLibrary extends Seed
      * @param string $id
      *            相簿id
      * @return array $allphoto
-     *         [
+     *     [Token =>[
      *         Title => Title,
      *         Description => Description,
      *         Path => Path
-     *         ];
-     *        
+     *     ]];
+     *         or
+     *         array $end = [ending => 'no album']
      */
     public static function allphoto($id)
     {
         $album = Album::load($id);
-        $aaa = $album->listPhoto();
-        foreach ($aaa as $a) {
-            $photo = Photo::load($a);
-            $allphoto[$a] = [
-                'Title' => $photo->getTitle(),
-                'Description' => $photo->getDescription(),
-                'Path' => 'aboard/photo/show/' . $photo->getToken()
+        if ($album != null) {
+            $aaa = $album->listPhoto();
+            foreach ($aaa as $a) {
+                $photo = Photo::load($a);
+                $allphoto[$a] = [
+                    'Title' => $photo->getTitle(),
+                    'Description' => $photo->getDescription(),
+                    'Path' => 'aboard/photo/show/' . $photo->getToken()
+                ];
+            }
+            return $allphoto;
+        } else {
+            $end = [
+                "ending" => "no album"
             ];
+            return $end;
         }
-        return $allphoto;
+    }
+
+    /**
+     * 取得album($id)的相簿資料
+     *
+     * @param string $id
+     *            相簿id
+     * @return array $albummsg
+     *         [
+     *         Title => Title,
+     *         Description => Description,
+     *         Token => Token,
+     *         create => create
+     *         ];
+     *         or
+     *         array $end = [ending => 'no album']
+     */
+    public static function albummsg($id)
+    {
+        $album = Album::load($id);
+        if ($album != null) {
+            $albummsg = [
+                'Title' => $album->getTitle(),
+                'Description' => $album->getDescription(),
+                'Token' => $album->getToken(),
+                'Create' => $album->getCreate()
+            ];
+        } else {
+            $end = [
+                "ending" => "no album"
+            ];
+            return $end;
+        }
     }
 
     /**
      * 取得album新增完成後資料
      *
-     * @param
-     *            string title 相簿名
-     * @param
-     *            string desc 相簿描述
-     * @return array $aaa
+     * @param string $title
+     *            相簿名
+     * @param string $desc
+     *            相簿描述
+     * @return array $end
      *         [
      *         Title => Title,
      *         Description => Description,
+     *         create => create,
      *         Token => Token
      *         ];
      */
@@ -80,12 +128,13 @@ class AlbumLibrary extends Seed
         $t = $title;
         $d = $desc;
         $album = Album::create($t, $d);
-        $aaa["0"] = [
+        $end["0"] = [
             "title" => $album->getTitle(),
             "desc" => $album->getDescription(),
-            "create" => $album->getCreate()
+            "create" => $album->getCreate(),
+            'Token' => $album->getToken()
         ];
-        return $aaa;
+        return $end;
     }
 
     /**
@@ -93,24 +142,31 @@ class AlbumLibrary extends Seed
      *
      * @param string $id
      *            相簿id
-     * @param
-     *            string title 相簿名
-     * @param
-     *            string desc 相簿描述
-     * @return array $end = [ending => 'edit okay']
+     * @param string $title
+     *            相簿名
+     * @param string $desc
+     *            相簿描述
+     * @return array $end = [ending => 'edit okay' or 'edit error']
      */
-    public static function edit($id)
+    public static function edit($id, $title, $desc)
     {
-        $t = $_POST["title"];
-        $d = $_POST["desc"];
+        $t = $title;
+        $d = $desc;
         $album = Album::load($id);
-        $album->setTitle($t);
-        $album->setDescription($d);
-        $album->save();
-        $end = [
-            "ending" => 'edit okay'
-        ];
-        return $end;
+        if ($album != null) {
+            $album->setTitle($t);
+            $album->setDescription($d);
+            $album->save();
+            $end = [
+                "ending" => 'edit okay'
+            ];
+            return $end;
+        } else {
+            $end = [
+                "ending" => "edit error"
+            ];
+            return $end;
+        }
     }
 
     /**
@@ -118,20 +174,28 @@ class AlbumLibrary extends Seed
      *
      * @param string $id
      *            相簿id
-     * @return array $end = [ending => 'delete okay']
+     * @return array $end = [ending => 'delete okay' or 'delete error']
      */
     public static function del($id)
     {
         $album = Album::load($id);
-        $aaa = $album->listPhoto();
-        foreach ($aaa as $a) {
-            $photo = Photo::load($a);
-            $photo->delete();
+        if ($album != null) {
+            $album->delete();
+            $end = [
+                "ending" => "delete okay"
+            ];
+            $aaa = $album->listPhoto();
+            foreach ($aaa as $a) {
+                $photo = Photo::load($a);
+                $photo->delete();
+                return $end;
+            }
+        } else {
+            $end = [
+                "ending" => "delete error"
+            ];
+            return $end;
         }
-        $album->delete();
-        $end = [
-            "ending" => 'delete okay'
-        ];
-        return $end;
     }
 }
+
